removed prescription photo module
