--- app/src/main/java/com/heartcare/agni/data/local/repository/file/DownloadedFileRepository.kt
@@ -1,9 +1,0 @@
-package com.heartcare.agni.data.local.repository.file
-
-import com.heartcare.agni.data.local.roomdb.entities.file.DownloadedFileEntity
-
-
-interface DownloadedFileRepository {
-    suspend fun insertEntity(downloadedFileEntity: DownloadedFileEntity): List<Long>
-    suspend fun getFileNames(): List<String>
-}
