+++ app/src/main/java/com/heartcare/agni/data/local/repository/prescription/PrescriptionRepository.kt
@@ -1,6 +1,5 @@
 package com.heartcare.agni.data.local.repository.prescription
 
-import com.heartcare.agni.data.local.model.prescription.PrescriptionPhotoResponseLocal
 import com.heartcare.agni.data.local.model.prescription.PrescriptionResponseLocal
 import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionAndMedicineRelation
 import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionDirectionsEntity
@@ -8,20 +7,8 @@ import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionDi
 interface PrescriptionRepository {
 
     suspend fun insertPrescription(prescriptionResponseLocal: PrescriptionResponseLocal): Long
-    suspend fun insertPhotoPrescription(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Long
     suspend fun getLastPrescription(patientId: String): List<PrescriptionResponseLocal>
     suspend fun getLastPrescriptionAndMedicine(patientId: String): List<PrescriptionAndMedicineRelation>
-    suspend fun getLastPhotoPrescription(patientId: String): List<PrescriptionPhotoResponseLocal>
     suspend fun getPrescriptionByAppointmentId(appointmentId: String): List<PrescriptionResponseLocal>
-    suspend fun getPrescriptionPhotoByAppointmentId(appointmentId: String): List<PrescriptionPhotoResponseLocal>
-    suspend fun getPrescriptionPhotoByDate(
-        patientId: String,
-        startDate: Long,
-        endDate: Long
-    ): PrescriptionPhotoResponseLocal
-    suspend fun getPrescriptionPhotoById(prescriptionId: String): PrescriptionPhotoResponseLocal
-
-    suspend fun insertPrescriptionPhotos(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Long
-    suspend fun deletePhotoPrescription(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Int
     suspend fun deletePrescriptionDirectionEntity(prescriptionDirectionsEntity: PrescriptionDirectionsEntity): Int
 }
