+++ app/src/main/java/com/heartcare/agni/data/local/repository/prescription/PrescriptionRepositoryImpl.kt
@@ -1,21 +1,16 @@
 package com.heartcare.agni.data.local.repository.prescription
 
-import com.heartcare.agni.data.local.model.prescription.PrescriptionPhotoResponseLocal
 import com.heartcare.agni.data.local.model.prescription.PrescriptionResponseLocal
-import com.heartcare.agni.data.local.roomdb.dao.FileUploadDao
 import com.heartcare.agni.data.local.roomdb.dao.PrescriptionDao
 import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionAndMedicineRelation
 import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionDirectionsEntity
 import com.heartcare.agni.utils.converters.responseconverter.toListOfPrescriptionDirectionsEntity
-import com.heartcare.agni.utils.converters.responseconverter.toListOfPrescriptionPhotoEntity
 import com.heartcare.agni.utils.converters.responseconverter.toPrescriptionEntity
-import com.heartcare.agni.utils.converters.responseconverter.toPrescriptionPhotoResponseLocal
 import com.heartcare.agni.utils.converters.responseconverter.toPrescriptionResponseLocal
 import javax.inject.Inject
 
 class PrescriptionRepositoryImpl @Inject constructor(
-    private val prescriptionDao: PrescriptionDao,
-    private val fileUploadDao: FileUploadDao
+    private val prescriptionDao: PrescriptionDao
 ) :
     PrescriptionRepository {
 
@@ -23,14 +18,6 @@ class PrescriptionRepositoryImpl @Inject constructor(
         return prescriptionDao.insertPrescription(prescriptionResponseLocal.toPrescriptionEntity())[0].also {
             prescriptionDao.insertPrescriptionMedicines(
                 *prescriptionResponseLocal.toListOfPrescriptionDirectionsEntity().toTypedArray()
-            )
-        }
-    }
-
-    override suspend fun insertPhotoPrescription(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Long {
-        return prescriptionDao.insertPrescription(prescriptionPhotoResponseLocal.toPrescriptionEntity())[0].also {
-            prescriptionDao.insertPrescriptionPhotos(
-                *prescriptionPhotoResponseLocal.toListOfPrescriptionPhotoEntity().toTypedArray()
             )
         }
     }
@@ -43,49 +30,11 @@ class PrescriptionRepositoryImpl @Inject constructor(
         return prescriptionDao.getPastPrescriptions(patientId)
     }
 
-    override suspend fun getLastPhotoPrescription(patientId: String): List<PrescriptionPhotoResponseLocal> {
-        return prescriptionDao.getPastPhotoPrescriptions(patientId).map { it.toPrescriptionPhotoResponseLocal() }
-    }
-
     override suspend fun getPrescriptionByAppointmentId(appointmentId: String): List<PrescriptionResponseLocal> {
         return prescriptionDao.getPrescriptionByAppointmentId(appointmentId)
             .map { prescriptionAndMedicineRelation ->
                 prescriptionAndMedicineRelation.toPrescriptionResponseLocal()
             }
-    }
-
-    override suspend fun getPrescriptionPhotoByAppointmentId(appointmentId: String): List<PrescriptionPhotoResponseLocal> {
-        return prescriptionDao.getPrescriptionPhotoByAppointmentId(appointmentId)
-            .map {
-                it.toPrescriptionPhotoResponseLocal()
-            }
-    }
-
-    override suspend fun getPrescriptionPhotoByDate(
-        patientId: String,
-        startDate: Long,
-        endDate: Long
-    ): PrescriptionPhotoResponseLocal {
-        return prescriptionDao.getPrescriptionPhotoByDate(patientId, startDate, endDate)
-            .map { it.toPrescriptionPhotoResponseLocal() }[0]
-    }
-
-    override suspend fun getPrescriptionPhotoById(prescriptionId: String): PrescriptionPhotoResponseLocal {
-        return prescriptionDao.getPrescriptionPhotoById(prescriptionId)
-            .toPrescriptionPhotoResponseLocal()
-    }
-
-    override suspend fun insertPrescriptionPhotos(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Long {
-        return prescriptionDao.insertPrescriptionPhotos(
-            *prescriptionPhotoResponseLocal.toListOfPrescriptionPhotoEntity().toTypedArray()
-        )[0]
-    }
-
-    override suspend fun deletePhotoPrescription(prescriptionPhotoResponseLocal: PrescriptionPhotoResponseLocal): Int {
-        fileUploadDao.deleteFile(prescriptionPhotoResponseLocal.prescription[0].filename)
-        return prescriptionDao.deletePrescriptionPhoto(prescriptionPhotoResponseLocal.toListOfPrescriptionPhotoEntity()[0]).also {
-            prescriptionDao.deletePrescriptionEntity(prescriptionPhotoResponseLocal.toPrescriptionEntity())
-        }
     }
 
     override suspend fun deletePrescriptionDirectionEntity(prescriptionDirectionsEntity: PrescriptionDirectionsEntity): Int {
