--- app/src/main/java/com/heartcare/agni/data/local/roomdb/entities/prescription/photo/PrescriptionAndFileEntity.kt
@@ -1,15 +1,0 @@
-package com.heartcare.agni.data.local.roomdb.entities.prescription.photo
-
-import androidx.annotation.Keep
-import androidx.room.Embedded
-import androidx.room.Relation
-import com.heartcare.agni.data.local.roomdb.entities.prescription.PrescriptionEntity
-
-@Keep
-data class PrescriptionAndFileEntity(
-    @Embedded val prescriptionEntity: PrescriptionEntity,
-    @Relation(
-        parentColumn = "id",
-        entityColumn = "prescriptionId"
-    ) val prescriptionPhotoEntity: List<PrescriptionPhotoEntity>
-)
