+++ app/src/main/java/com/heartcare/agni/data/server/api/PrescriptionApiService.kt
@@ -5,13 +5,10 @@ import com.heartcare.agni.data.server.constants.EndPoints.MEDICATION_REQUEST
 import com.heartcare.agni.data.server.model.create.CreateResponse
 import com.heartcare.agni.data.server.model.prescription.medication.MedicationResponse
 import com.heartcare.agni.data.server.model.prescription.medication.MedicineTimeResponse
-import com.heartcare.agni.data.server.model.prescription.photo.PrescriptionPhotoResponse
 import com.heartcare.agni.data.server.model.prescription.prescriptionresponse.PrescriptionResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.HTTP
-import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Path
@@ -32,17 +29,8 @@ interface PrescriptionApiService {
     @GET("Prescription")
     suspend fun getPastPrescription(@QueryMap(encoded = true) map: Map<String, String>?): Response<BaseResponse<List<PrescriptionResponse>>>
 
-    @GET("PrescriptionFile")
-    suspend fun getPastPhotoPrescription(@QueryMap(encoded = true) map: Map<String, String>?): Response<BaseResponse<List<PrescriptionPhotoResponse>>>
-
     @GET("sct/medTime")
     suspend fun getMedicineTime(@QueryMap(encoded = true) map: Map<String, String>?): Response<BaseResponse<List<MedicineTimeResponse>>>
-
-    @PATCH("DocumentReference")
-    suspend fun patchListOfChanges(@Body patchLogs: List<Any>): Response<BaseResponse<List<CreateResponse>>>
-
-    @HTTP(method = "DELETE", path = "PrescriptionFile", hasBody = true)
-    suspend fun deletePrescriptionPhoto(@Body patchLogs: List<Any>): Response<BaseResponse<List<CreateResponse>>>
 
     @PUT(MEDICATION_REQUEST)
     suspend fun sendPrescriptionPut(@Body prescriptionResponses: List<PrescriptionResponse>): Response<BaseResponse<List<CreateResponse>>>
