--- app/src/main/java/com/heartcare/agni/data/server/model/file/request/FilesRequest.kt
@@ -1,8 +1,0 @@
-package com.heartcare.agni.data.server.model.file.request
-
-import androidx.annotation.Keep
-
-@Keep
-data class FilesRequest(
-    val files: List<String>
-)
