--- app/src/main/java/com/heartcare/agni/data/server/model/file/response/FilesResponse.kt
@@ -1,17 +1,0 @@
-package com.heartcare.agni.data.server.model.file.response
-
-import androidx.annotation.Keep
-
-@Keep
-data class FilesResponse(
-    val files: List<Files>,
-    val errors: List<Files>
-)
-
-@Keep
-data class Files(
-    val originalName: String,
-    val filename: String,
-    val url: String?,
-    val error: String?
-)
