--- app/src/main/java/com/heartcare/agni/data/server/model/prescription/medication/Strength.kt
@@ -1,10 +1,0 @@
-package com.heartcare.agni.data.server.model.prescription.medication
-
-import androidx.annotation.Keep
-
-@Keep
-data class Strength(
-    val medName: String,
-    val unitMeasureValue: String,
-    val medMeasureCode: String
-)
