--- app/src/main/java/com/heartcare/agni/data/server/model/prescription/photo/PrescriptionPhotoPatch.kt
@@ -1,10 +1,0 @@
-package com.heartcare.agni.data.server.model.prescription.photo
-
-import androidx.annotation.Keep
-
-@Keep
-data class PrescriptionPhotoPatch(
-    val documentFhirId: String,
-    val note: String,
-    val filename: String
-)
