--- app/src/main/java/com/heartcare/agni/data/server/model/prescription/photo/PrescriptionPhotoResponse.kt
@@ -1,20 +1,0 @@
-package com.heartcare.agni.data.server.model.prescription.photo
-
-import androidx.annotation.Keep
-import com.google.gson.annotations.SerializedName
-import java.util.Date
-
-@Keep
-data class PrescriptionPhotoResponse(
-    val appointmentUuid: String,
-    val prescriptionId: String,
-    @SerializedName("prescriptionDocumentFhirId")
-    val prescriptionFhirId: String?,
-    val appointmentId: String,
-    @SerializedName("patientId")
-    val patientFhirId: String,
-    val generatedOn: Date,
-    @SerializedName("prescriptionFiles")
-    val prescription: List<File>,
-    val status: String?
-)
