--- app/src/main/java/com/heartcare/agni/ui/prescription/model/PrescriptionFormAndPhoto.kt
@@ -1,9 +1,0 @@
-package com.heartcare.agni.ui.prescription.model
-
-import java.util.Date
-
-data class PrescriptionFormAndPhoto(
-    val date: Date,
-    val type: String,
-    val prescription: Any
-)
